backend/src/social/chat_service.py
import logging
from typing import Dict, Any, List
from sqlalchemy.orm import Session
from ..database.models import League, Team, User
import json
import uuid
from datetime import datetime

class LeagueChatService:
    """
    League chat service that provides messaging functionality for fantasy football leagues.
    This service allows users to send and receive messages within their leagues.
    """
    
    def __init__(self, db_session: Session = None):
        """Initialize the league chat service."""
        self.db_session = db_session
        self.service_version = "1.0"
        
    def send_message(self, league_id: str, user_id: str, message_content: str) -> Dict[str, Any]:
        """
        Send a message to a league chat.
        
        Args:
            league_id (str): League ID to send message to
            user_id (str): User ID of the sender
            message_content (str): Content of the message
            
        Returns:
            dict: Message details including ID, timestamp, and status
        """
        if not self.db_session:
            logging.warning("No database session provided, cannot send message")
            return {"status": "error", "message": "No database session provided"}
            
        try:
            # Verify that the user is part of the league
            league = self.db_session.query(League).filter(League.id == league_id).first()
            if not league:
                return {"status": "error", "message": "League not found"}
                
            # Create a new message (in a real implementation, this would be a database model)
            message_id = str(uuid.uuid4())
            timestamp = datetime.utcnow()
            
            message = {
                "id": message_id,
                "league_id": league_id,
                "user_id": user_id,
                "content": message_content,
                "timestamp": timestamp.isoformat(),
                "status": "sent"
            }
            
            # In a real implementation, we would save this to a database table
            # For now, we'll just log it
            logging.info(f"Message sent to league {league_id}: {message_content}")
            
            return message
            
        except Exception as e:
            logging.error(f"Error sending message to league {league_id}: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    def get_league_messages(self, league_id: str, limit: int = 50) -> List[Dict[str, Any]]:
        """
        Get messages for a league chat.
        
        Args:
            league_id (str): League ID to get messages for
            limit (int): Maximum number of messages to retrieve (default: 50)
            
        Returns:
            list: List of messages in the league chat
        """
        if not self.db_session:
            logging.warning("No database session provided, returning empty message list")
            return []
            
        try:
            # Verify that the league exists
            league = self.db_session.query(League).filter(League.id == league_id).first()
            if not league:
                return []
                
            # In a real implementation, this would query a messages table
            # For now, we'll return an empty list
            messages = []
            
            return messages
            
        except Exception as e:
            logging.error(f"Error getting messages for league {league_id}: {str(e)}")
            return []
    
    def get_user_league_chats(self, user_id: str) -> List[Dict[str, Any]]:
        """
        Get all league chats for a user.
        
        Args:
            user_id (str): User ID to get league chats for
            
        Returns:
            list: List of league chats the user is part of
        """
        if not self.db_session:
            logging.warning("No database session provided, returning empty league chat list")
            return []
            
        try:
            # Get all leagues for the user
            leagues_query = self.db_session.query(League).filter(League.user_id == user_id)
            leagues = leagues_query.all()
            
            league_chats = []
            for league in leagues:
                # Get teams in the league to show chat participants
                teams_query = self.db_session.query(Team).filter(Team.league_id == league.id)
                teams = teams_query.all()
                participants = []
                for team in teams:
                    participants.append({
                        "id": team.id,
                        "name": team.team_name
                    })
                
                league_chats.append({
                    "league_id": league.id,
                    "league_name": getattr(league, 'league_name', 'Unknown League'),
                    "platform": getattr(league, 'platform', 'Unknown Platform'),
                    "participants": participants,
                    "message_count": 0,  # In a real implementation, this would be a count of messages
                    "last_message_timestamp": None
                })
                
            return league_chats
            
        except Exception as e:
            logging.error(f"Error getting league chats for user {user_id}: {str(e)}")
            return []
