backend/requirements.txt
# Backend Python Dependencies

# Web Framework
fastapi==0.68.0
uvicorn==0.15.0

# Database
psycopg2-binary==2.9.10
redis==3.5.3

# Data Processing
pandas==1.3.2
numpy==1.21.2

# Machine Learning
scikit-learn==0.24.2
tensorflow==2.6.0

# API Clients
requests==2.26.0
firecrawl-py==1.0.0

# Security
cryptography==3.4.8

# Streaming
kafka-python==2.0.2

# Utilities
pydantic==1.8.2
python-dotenv==0.18.0
celery==5.1.2

# Testing
pytest==6.2.4
pytest-cov==2.12.1
